Details in the description file of the functions:
I added detailled information in the functions' description files. For all of the functions except hetest I provided a @return (Value in the output) in the roxygen2 description.

I checked All the example provided in each description file and they rrun perfectely.

I checked the result from the fixed effect model and the random effect model on the tuberculosis dataset and the results provided are perfectely equal to those provided in the stata book.